Document LISTEN_FDS for passing open file descriptors
This landed in runC with [1], but the bundle-author <-> runtime specs
explicitly avoid talking about how this is set (since the
bundle-author doesn't care about the runtime-caller <-> runtime
interface) [2].  However, *this* spec is about the runtime-caller <->
runtime interface, so we need to document it here.

I've left LISTEN_PID [3,4] out, since I don't see how the
runtime-caller would choose anything other than 1 for its value.  It
seems like something that a process would have to set for itself
(because guessing the PID of a child before spawning it seems racy ;).
In any event, the runC implementation seems to set this to 1
regardless of what systemd passes to it [4].

I've borrowed Shishir's wording for the example [4].

[1]: https://github.com/opencontainers/runc/pull/231
[2]: https://github.com/opencontainers/specs/pull/113#issuecomment-135857404
[3]: http://www.freedesktop.org/software/systemd/man/sd_listen_fds.html
[4]: https://github.com/opencontainers/runc/pull/231#issuecomment-135469477

Signed-off-by: W. Trevor King <wking@tremily.us>